+++ ldap4j-java/src/main/java/hu/gds/ldap4j/ldap/Control.java
@@ -1,6 +1,8 @@
 package hu.gds.ldap4j.ldap;
 
+import hu.gds.ldap4j.Function;
 import hu.gds.ldap4j.net.ByteBuffer;
+import java.util.List;
 import java.util.Objects;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -16,6 +18,21 @@ public record Control(
         this.controlType=Objects.requireNonNull(controlType, "controlType");
         this.controlValue=controlValue;
         this.criticality=criticality;
+    }
+
+    public static <T> @Nullable T findOne(
+            @NotNull List<@NotNull Control> controls,
+            @NotNull Function<@NotNull Control, T> function)
+            throws Throwable {
+        @Nullable T result=null;
+        for (@NotNull Control control: controls) {
+            @Nullable T value=function.apply(control);
+            if (null!=result) {
+                throw new RuntimeException("more than one matching control");
+            }
+            result=value;
+        }
+        return result;
     }
 
     public static Control nonCritical(@NotNull String controlType) {