+++ ldap4j-java/src/main/java/hu/gds/ldap4j/ldap/ExtendedRequest.java
@@ -46,6 +46,9 @@ public record ExtendedRequest(
             Ldap.EXTENDED_REQUEST_FAST_BIND_OID, null, ExtendedResponse.READER_SUCCESS);
     public static final @NotNull ExtendedRequest START_TLS=new ExtendedRequest(
             Ldap.EXTENDED_REQUEST_START_TLS_OID, null, ExtendedResponse.READER_SUCCESS);
+    /**
+     * RFC 4532
+     */
     public static final @NotNull ExtendedRequest WHO_AM_I=new ExtendedRequest(
             Ldap.EXTENDED_REQUEST_WHO_AM_I, null, ExtendedResponse.READER_SUCCESS);
 
@@ -58,6 +61,9 @@ public record ExtendedRequest(
         this.responseReader=Objects.requireNonNull(responseReader, "responseReader");
     }
     
+    /**
+     * RFC 3909
+     */
     public static @NotNull ExtendedRequest cancel(int messageId) {
         return new ExtendedRequest(
                 Ldap.EXTENDED_REQUEST_CANCEL_OP_OID,